backend/requirements.txt
# Django Backend Requirements
django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
firebase-admin>=6.4.0
python-decouple>=3.8
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
pillow>=10.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
google-generativeai>=0.3.0
